micro-service-server/requirements.txt
aioredis
starlette
httpx
fastapi
uvicorn[standard]
sqlalchemy[asyncio]>=2.0.0
pydantic
pydantic-settings
python-dotenv
alembic
asyncpg>=0.28.0
python-jose[cryptography]
redis>=4.5.0,<5.0.0
python-multipart